requirements.txt
# Web Framework & Server
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
gunicorn>=23.0.0
Jinja2>=3.1.0
python-multipart>=0.0.20

# Database & Task Queue
SQLAlchemy>=2.0.49
huey>=2.5.0

# Configuration & Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.2
httpx>=0.28.0
pydantic>=2.13.0
pydantic-settings>=2.0.0
werkzeug>=3.0.0
itsdangerous>=2.2.0
authlib>=1.4.0
bleach>=6.2.0

# PDF Processing (primary)
pymupdf>=1.25.0

# OCR & Image Processing
pytesseract>=0.3.13
Pillow>=11.0.0
opencv-python-headless>=4.10.0

# Transcription (optional, large models)
faster-whisper>=1.1.0

# TTS (optional)
piper-tts
kokoro-tts

# Document conversion helpers
markitdown[pdf,docx,pptx,xlsx,outlook]>=0.1.3
